requirements.txt
fastapi
jinja2
pdfplumber
python-docx
xlrd
uvicorn
watchdog
requests
openai
python-multipart
pandas
openpyxl
langchain-huggingface
chromadb
langchain-community
langchain-chroma
langchain-text-splitters
sentence-transformers
